requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.4.3
scikit-image>=0.18.3
transformers>=4.11.0
pillow>=8.3.2 